Moved spatial_processing into app/sofast/lib
