+++ opencsp/common/lib/deflectometry/process_optics_geometry.py
@@ -16,7 +16,7 @@ from opencsp.common.lib.deflectometry.GeometryProcessingParams import (
 from opencsp.common.lib.deflectometry.GeometryDataDebug import GeometryDataDebug
 import opencsp.common.lib.deflectometry.image_processing as ip
 from opencsp.app.sofast.lib.SpatialOrientation import SpatialOrientation
-import opencsp.common.lib.deflectometry.spatial_processing as sp
+import opencsp.app.sofast.lib.spatial_processing as sp
 from opencsp.common.lib.geometry.LoopXY import LoopXY
 from opencsp.common.lib.geometry.Uxyz import Uxyz
 from opencsp.common.lib.geometry.Vxy import Vxy
